move EncryptRuleAware to rewrite package
